Split templates in seperate file
