--- packages/mux-player/src/env.js
@@ -1,12 +1,0 @@
-/* eslint-disable */
-const getEnvPlayerVersion = () => {
-  try {
-    // @ts-ignore
-    return PLAYER_VERSION;
-  } catch {}
-  return "UNKNOWN";
-};
-
-const player_version = getEnvPlayerVersion();
-
-export const getPlayerVersion = () => player_version;
